refactor: change ID types from Long to Integer across all entities
- Update all entity primary key fields from Long to Integer to match
  database schema (INT columns), except analysis_id which remains Long
  as it maps to a BIGINT column in the database

- Files updated across 6 layers:
  * entity: Crop, GrowthProcess, InventoryBatch, InventoryItem,
    PlantingBatch, Role, Stage, Task, TaskTemplate, WeatherAlert
  * dto: All DTOs updated to match entity ID types
  * mapper: All mappers updated to be consistent with new types
  * repository: JpaRepository<Entity, Long> -> JpaRepository<Entity, Integer>
  * service: All service method signatures updated (findById, update,
    deleteById parameters changed to Integer)
  * controller: @PathVariable and @RequestParam changed from Long to Integer

- AiAnalysis is a special case: analysisId stays Long (BIGINT in DB),
  only the FK reference pBatchId is changed to Integer

- Also implemented previously empty service classes:
  AiAnalysisService, InventoryBatchService, StageService,
  TaskService, TaskTemplateService, UserService

+++ Server/farm-manager-api/src/main/java/com/smartfarm/api/controller/AiAnalysisController.java
@@ -24,7 +24,7 @@ public class AiAnalysisController {
 
     @GetMapping
     public ResponseEntity<ApiResponse<List<AiAnalysisDto>>> getAll(
-            @RequestParam(required = false) Long pBatchId) {
+            @RequestParam(required = false) Integer pBatchId) {
         try {
             List<AiAnalysisDto> data = (pBatchId != null)
                     ? aiAnalysisService.findByBatchId(pBatchId)
