+++ Server/farm-manager-api/src/main/java/com/smartfarm/api/dto/InventoryItemDto.java
@@ -7,11 +7,11 @@ import lombok.*;
 @NoArgsConstructor
 @AllArgsConstructor
 public class InventoryItemDto {
-    private Long itemId;
+    private Integer itemId;
     private String itemName;
     private String category;
     private String unit;
     private Double minThreshold;
-    private Long cropId;
+    private Integer cropId;
     private String cropName;
 }
