+++ Server/farm-manager-api/src/main/java/com/smartfarm/api/dto/TaskDto.java
@@ -8,15 +8,15 @@ import java.time.LocalDate;
 @NoArgsConstructor
 @AllArgsConstructor
 public class TaskDto {
-    private Long taskId;
-    private Long pBatchId;
+    private Integer taskId;
+    private Integer pBatchId;
     private String batchName;
-    private Long taskTmpId;
+    private Integer taskTmpId;
     private String taskTemplateName;
     private String title;
     private LocalDate plannedDate;
     private LocalDate actualDate;
-    private Long assignedToId;
+    private Integer assignedToId;
     private String assignedToName;
     private String status;
     private String notes;
