+++ Server/farm-manager-api/src/main/java/com/smartfarm/api/repository/InventoryItemRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface InventoryItemRepository extends JpaRepository<InventoryItem, Long> {
+public interface InventoryItemRepository extends JpaRepository<InventoryItem, Integer> {
     List<InventoryItem> findByCategory(String category);
 }
